Add Anthropic & utility to get LM in examples
