+++ core/examples/structured_data_generation_blog.rs
@@ -7,6 +7,12 @@ use orch::execution::*;
 use orch::lm::*;
 use orch::response::*;
 
+mod example_utils;
+use example_utils::get_lm;
+
+// ! Change this to use a different provider.
+pub const PROVIDER: LanguageModelProvider = LanguageModelProvider::Ollama;
+
 #[derive(Variants, serde::Deserialize)]
 #[serde(tag = "response_type")]
 pub enum ResponseVariants {
@@ -44,8 +50,7 @@ pub struct FailResponseVariant {
 
 #[tokio::main]
 async fn main() {
-    // ! Change this to use a different provider.
-    let provider = LanguageModelProvider::OpenAi;
+    let lm = get_lm(PROVIDER);
 
     let args = std::env::args().collect::<Vec<_>>();
     let blog_file_path = args.get(1).unwrap_or_else(|| {
@@ -55,30 +60,6 @@ async fn main() {
     let prompt = std::fs::read_to_string(blog_file_path).expect("Failed to read blog file");
 
     println!("Analyzing blog post at path '{blog_file_path}'...");
-
-    // Use a different language model, per the `provider` variable (feel free to change it).
-    let open_ai_api_key = {
-        if provider == LanguageModelProvider::OpenAi {
-            std::env::var("OPENAI_API_KEY")
-                .unwrap_or_else(|_| panic!("OPENAI_API_KEY environment variable not set"))
-        } else {
-            String::new()
-        }
-    };
-    let lm: Box<dyn LanguageModel> = match provider {
-        LanguageModelProvider::Ollama => Box::new(
-            OllamaBuilder::new()
-                .with_model(ollama_model::PHI3_MINI.to_string())
-                .try_build()
-                .unwrap(),
-        ),
-        LanguageModelProvider::OpenAi => Box::new(
-            OpenAiBuilder::new()
-                .with_api_key(open_ai_api_key)
-                .try_build()
-                .unwrap(),
-        ),
-    };
 
     let executor = StructuredExecutorBuilder::new()
         .with_lm(&*lm)
