+++ core/examples/structured_data_generation_capital.rs
@@ -7,6 +7,12 @@ use orch::execution::*;
 use orch::lm::*;
 use orch::response::*;
 
+mod example_utils;
+use example_utils::get_lm;
+
+// ! Change this to use a different provider.
+pub const PROVIDER: LanguageModelProvider = LanguageModelProvider::Ollama;
+
 #[derive(Variants, serde::Deserialize)]
 pub enum ResponseVariants {
     Answer(AnswerResponseVariant),
@@ -43,33 +49,13 @@ pub struct FailResponseVariant {
 
 #[tokio::main]
 async fn main() {
-    // ! Change this to use a different provider.
-    let provider = LanguageModelProvider::Ollama;
+    let lm = get_lm(PROVIDER);
 
     let args = std::env::args().collect::<Vec<_>>();
     let prompt = args.get(1).unwrap_or_else(|| {
         eprintln!("ERROR: Please provide a country name");
         std::process::exit(1);
     });
-
-    // Use a different language model, per the `provider` variable (feel free to change it).
-    let open_ai_api_key = {
-        if provider == LanguageModelProvider::OpenAi {
-            std::env::var("OPENAI_API_KEY")
-                .unwrap_or_else(|_| panic!("OPENAI_API_KEY environment variable not set"))
-        } else {
-            String::new()
-        }
-    };
-    let lm: Box<dyn LanguageModel> = match provider {
-        LanguageModelProvider::Ollama => Box::new(OllamaBuilder::new().try_build().unwrap()),
-        LanguageModelProvider::OpenAi => Box::new(
-            OpenAiBuilder::new()
-                .with_api_key(open_ai_api_key)
-                .try_build()
-                .unwrap(),
-        ),
-    };
 
     let executor = StructuredExecutorBuilder::new()
         .with_lm(&*lm)
